split up monitor into submodules to allow for missing matplotlib import
