+++ gpflow/monitor/__init__.py
@@ -1,0 +1,25 @@
+# Copyright 2020 GPflow authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+""" Provides basic functionality to monitor optimisation runs """
+
+from .base import *
+from .tensorboard import *
+
+try:
+    from .image_to_tensorboard import ImageToTensorBoard
+except ImportError:
+
+    class ImageToTensorBoard(ToTensorBoard):
+        def __init__(self, *args, **kwargs):
+            raise NotImplementedError("ImageToTensorBoard requires matplotlib to be installed")
